Rename dialogs module to builtin_dialogs, fix wrong import order of packages and modules, fix module comments
