+++ src/main/python/playlist/video_info.py
@@ -2,9 +2,9 @@
 
 import logging
 
-from time_and_date import convert_time
-
 from pytube import Playlist, YouTube, exceptions
+
+from time_and_date import convert_time
 
 
 def get_video_info(video_id: str) -> dict | None:
