+++ src/main/python/settings/operations.py
@@ -1,4 +1,4 @@
-"""settings.operations module"""
+"""module settings.operations"""
 
 import json
