+++ src/main/python/settings/settings.py
@@ -2,10 +2,10 @@
 
 APP_VERSION = "0.2.0"
 APP_ICON = "icon/youtube-play.icns"
-RELEASE_DATE = "Nov 30 2022"
-RECENT_FILES_STRING = "Clear recent files"
-SETTING_FILE_LOCATION = "config/settings.config"
-DEFAULT_SETTINGS_FILE_LOCATION = "config/settings_default.config"
 DARK_THEME_LOCATION = "theme/yt-dark-red.xml"
-WHITE_THEME_LOCATION = "theme/yt-white-red.xml"
+DEFAULT_SETTINGS_FILE_LOCATION = "config/settings_default.config"
 FILE_NOT_FOUND_STRING = "File not found. No file was imported."
+RECENT_FILES_STRING = "Clear recent files"
+RELEASE_DATE = "Nov 30 2022"
+SETTING_FILE_LOCATION = "config/settings.config"
+WHITE_THEME_LOCATION = "theme/yt-white-red.xml"
